fix: consolidate gaming integration module structure
