+++ src/gaming/gaming_integration_core.py
@@ -1,5 +1,5 @@
-"""Backward compatible stub for :mod:`integration_core`."""
+"""Backward compatible stub for :mod:`gaming.integration.core`."""
 
-from .integration_core import GamingIntegrationCore, GameType, IntegrationStatus
+from .integration.core import GamingIntegrationCore, GameType, IntegrationStatus
 
 __all__ = ["GamingIntegrationCore", "GameType", "IntegrationStatus"]
