+++ src/gaming/integration/__init__.py
@@ -1,0 +1,12 @@
+"""Gaming integration package."""
+
+from .core import GamingIntegrationCore, GameType, IntegrationStatus
+from .models import GameSession, EntertainmentSystem
+
+__all__ = [
+    "GamingIntegrationCore",
+    "GameType",
+    "IntegrationStatus",
+    "GameSession",
+    "EntertainmentSystem",
+]
